+++ include/counting.h
@@ -8,8 +8,8 @@
 
 using namespace std;
 
-void CountingLetters(ifstream &flux, string &RawText, int &LineCnt, int *LetterCount[]);
-void CountingNumbers(ifstream &flux, string &RawText, int &LineCnt, int *NumbersCount[]);
-void CountingAll(ifstream &flux, string &RawText, int &LineCnt, vector<char> &CharList, vector<int> &FileCount);
+void CountingLetters(ifstream &flux, string &RawText, long int &LineCnt, long int *LetterCount[]);
+void CountingNumbers(ifstream &flux, string &RawText, long int &LineCnt, long int *NumbersCount[]);
+void CountingAll(ifstream &flux, string &RawText, long int &LineCnt, vector<char> &CharList, vector<long int> &FileCount);
 
 #endif
